(feat) Use consistently denoramlised environment
This does share code with the activity-stream. However, we do not factor it
out, since we don't actually expect to use this repo. It's only here for
just-in-case we do want a stand alone proxy in the future, so we can pen test
it now and be to quickly use it it later if we need to. Slightly un-agile, but
the cost of pen testing later is high, both in terms of money and time.